fix a bug with created nodes not being added to document
The new implementation that kept references live by using a
DocumentFragment had a bug because it iterated over a live NodeList and
added its members to the DocumentFragment, which in turn would remove
them from the _live_ NodeList. This lead to odd-number-indexed items
(starting from 0) to be skipped in the loop. Fixed by first "freezing"
the NodeList with Array.from().